Add hint action table.
