--- hanabi/models/__init__.py
@@ -1,7 +1,0 @@
-from .action import Action
-from .card import Card
-from .game import Game
-from .lobby import Lobby
-from .player import Player
-
-__all__ = ["Action", "Card", "Game", "Lobby", "Player"]
